Moved libs directory to common.
