+++ prototype2/common/BitMath.h
