+++ prototype2/common/HistSerializer.cpp
@@ -1,7 +1,7 @@
 /** Copyright (C) 2016, 2017 European Spallation Source ERIC */
 
 #include <common/HistSerializer.h>
-#include <libs/include/gccintel.h>
+#include <common/gccintel.h>
 
 #include <common/Trace.h>
