+++ prototype2/common/ReadoutSerializer.cpp
@@ -3,7 +3,7 @@
 #include <cinttypes>
 #include <common/ReadoutSerializer.h>
 #include <common/Trace.h>
-#include <libs/include/gccintel.h>
+#include <common/gccintel.h>
 
 //#undef TRC_LEVEL
 //#define TRC_LEVEL TRC_L_DEB
