+++ prototype2/common/gccintel.h
