+++ prototype2/multiblade/MBCaenBase.h
@@ -11,7 +11,7 @@
 
 #include <common/Detector.h>
 #include <common/RingBuffer.h>
-#include <libs/include/SPSCFifo.h>
+#include <common/SPSCFifo.h>
 #include <caen/Config.h>
 #include <prototype2/multiblade/caen/Readout.h>
