+++ prototype2/readout/Readout.cpp
@@ -1,7 +1,7 @@
 /** Copyright (C) 2016, 2017 European Spallation Source ERIC */
 
 #include <common/Trace.h>
-#include <libs/include/gccintel.h> // UNUSED macros
+#include <common/gccintel.h> // UNUSED macros
 #include <readout/Readout.h>
 
 #define CKSUMSIZE 4U
