+++ prototype2/sonde/SoNDeBase.h
@@ -11,7 +11,7 @@
 #include <cinttypes>
 #include <common/Detector.h>
 #include <common/RingBuffer.h>
-#include <libs/include/SPSCFifo.h>
+#include <common/SPSCFifo.h>
 
 //#undef TRC_LEVEL
 //#define TRC_LEVEL TRC_L_DEB
