+++ prototype2/tools/udpgen_pcap.cpp
@@ -5,7 +5,7 @@
 #include <cinttypes>
 #include <tools/PcapArgs.h>
 #include <tools/ReaderPcap.h>
-#include <libs/include/Socket.h>
+#include <common/Socket.h>
 #include <string.h>
 #include <string>
 #include <unistd.h>
